Booking Receipt Prototype Finished
Still needs a bit of work implementing the charges, and the lists in particular do not look great with the rest of the report. However all of the current input is now mapped to the receipt. Besides Refining, next steps will be to add a finalize booking button to "Send" data to server and clear current data, and after which a button which can redirect you to the home menu or to create another booking